Fixed an error with an int return calling a double. Readded the halstead to the display

+++ src/sample/DataCollected.java
@@ -181,7 +181,7 @@ public class DataCollected {
 
     public int getCyclomaticComplexity() { return cyclomaticcomplexity; }
 
-    public int getMaintainabilityIndex() { return maintainabilityindex; }
+    public double getMaintainabilityIndex() { return maintainabilityindex; }
 
     public int getDepthOfInheritance() { return depthofinheritance; }
 }
